match_pathname(): drop unused "flags" parameter
This field has not been used since the function was introduced in
934d182b9a (exclude: split pathname matching code into a separate
function, 2012-10-15), though there was a brief period where it was
erroneously used and then reverted in e541ea1675 (dir: fix pattern
matching on dirs, 2021-09-24) and 2fe88d43c2 (dir: fix
directory-matching bug, 2021-11-02).

It's possible we'd eventually add a flag that makes it useful here, but
there are only a handful of callers. It would be easy to add back if
necessary, and in the meantime this makes the function interface less
misleading.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>